Tag import vendor 3.6

+++ caPutLogApp/caPutLogTask.c
@@ -152,7 +152,7 @@ int epicsShareAPI caPutLogTaskStart(int config)
     shut_down = FALSE;
     threadId = epicsThreadCreate("caPutLog", epicsThreadPriorityLow,
         epicsThreadGetStackSize(epicsThreadStackSmall),
-        caPutLogTask, (void*)config);
+        caPutLogTask, (void*)(size_t)config);
     if (!threadId) {
         errlogSevPrintf(errlogFatal,"caPutLog: thread creation failed\n");
         return caPutLogError;
@@ -178,7 +178,7 @@ void epicsShareAPI caPutLogTaskSend(LOGDATA *plogData)
 static void caPutLogTask(void *arg)
 {
     int sent = FALSE, burst = FALSE;
-    int config = (int)arg;
+    int config = (size_t)arg;
     LOGDATA *pcurrent, *pnext;
     VALUE old_value, max_value, min_value;
     VALUE *pold=&old_value, *pmax=&max_value, *pmin=&min_value;
